% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addColumns}
\alias{addColumns}
\title{Add header columns}
\usage{
addColumns(L, newnames)
}
\arguments{
\item{L}{a list of spectra}

\item{newnames}{a character vector holding an array of column names to be
added to the header}
}
\description{
Add new columns to the header part of the spectra, you may fill these new
columns by using routine 'modify'.
}
\examples{
data(salsa)
print(getHead(salsa))
addColumns(salsa, "T.sys")
modify(salsa, "T.sys", rep(300.0, 29))
print(getHead(salsa))    # header should now have additional column T.sys
}
\seealso{
\code{\link{modify}}
}

