% Generated by roxygen2 (4.0.1): do not edit by hand
\name{area}
\alias{area}
\title{Calculate integrated area}
\usage{
area(S, mask)
}
\arguments{
\item{S}{a single spectrum}

\item{mask}{a logical vector equal to TRUE for all the channels that should
be integrated.}
}
\value{
the integrated value
}
\description{
Given a mask defining the spectral areas of interest, return the integrated
area over those region(s).
}

