% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rdrp.R
\name{average}
\alias{average}
\title{Average spectra in a dataset}
\usage{
average(ds)
}
\arguments{
\item{ds}{a dataset}
}
\value{
a new dataset with the averaged spectrum as the only entry
}
\description{
Calculate a weighted sum of all spectra in a dataset.
}

