% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rdrp.R
\name{baseline}
\alias{baseline}
\title{Fit a baseline to spectra in a dataset}
\usage{
baseline(s, order = 1)
}
\arguments{
\item{s}{a single spectrum}

\item{order}{the order of the polynomial to fit}
}
\value{
the fitted baselines
}
\description{
For all spectra in a dataset fit a polynomial of given order, using only the
spectral channels given in window.
}

