% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rdrp.R
\name{boxcar}
\alias{boxcar}
\title{Apply boxcar smoothing.}
\usage{
boxcar(s, n = 3)
}
\arguments{
\item{s}{a single spectrum}

\item{n}{the length of the boxcar filter}
}
\value{
the filtered data vector
}
\description{
Filter the data vector by a a boxcar filter function.
}

