% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clip}
\alias{clip}
\title{Clip channels from spectra}
\usage{
clip(S, keep)
}
\arguments{
\item{S}{a single spectrum}

\item{keep}{a vector holding the channel numbers to keep}
}
\value{
the clipped spectrum
}
\description{
Given a single spectrum and a vector of channel numbers, clip the spectrum such that
only the given channels are kept in the frequency and data vectors.
}

