% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rdrp.R
\name{clipChannels}
\alias{clipChannels}
\title{Clip channels from spectra}
\usage{
clipChannels(ds, keep)
}
\arguments{
\item{ds}{a datase}

\item{keep}{an integer vector listing channel numbers to preserve}
}
\value{
a new dataset with fewer channels
}
\description{
Given a vector of channel numbers, only keep those channels from both the frequency
and data matrices.
}

