% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/drp.R
\name{examine}
\alias{examine}
\title{Examine a dataset or single spectrum}
\usage{
examine(s)
}
\arguments{
\item{s}{a dataset (class 'spectra') or a single spectrum (class 'spectrum')}
}
\description{
Print class and dimension information.
}

