% Generated by roxygen2 (4.0.1): do not edit by hand
\name{frequencies}
\alias{frequencies}
\title{Construct frequency vector}
\usage{
frequencies(S, v)
}
\arguments{
\item{S}{a single spectrum}

\item{v}{a numeric vector holding velocities}
}
\value{
a numeric vector holding frequencies
}
\description{
Turn a given velocity vector into a frequency vector.
}
\examples{
# resample spectrum at particular velocities
v <- seq(-100,100,by=2)
R <- resample(S, frequencies(S,v))
}
\seealso{
\code{\link{velocities}}
}

