% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeDataset.R
\name{getData}
\alias{getData}
\title{Return data part of spectra.}
\usage{
getData(L)
}
\arguments{
\item{L}{list of spectra}
}
\value{
a matrix containing one data vector per column.
}
\description{
Turn data component of list of spectra into matrix.
}

