% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getDimension}
\alias{getDimension}
\title{Get dimensions}
\usage{
getDimension(L)
}
\arguments{
\item{L}{a list of spectra, each with components 'head', 'freq' and 'data'}
}
\value{
a two component integer vector (nChannels, nSpectra)
}
\description{
From a list of spectra, get the dimensions of the data.
}

