% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeDataset.R
\name{getFreq}
\alias{getFreq}
\title{Return frequencies of spectra.}
\usage{
getFreq(L)
}
\arguments{
\item{L}{list of spectra}
}
\value{
a matrix containing one frequency vector per column.
}
\description{
Turn freq component of list of spectra into matrix.
}

