% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getOdinSpectrum}
\alias{getOdinSpectrum}
\title{Get a spectrum from an Odin binary file on disk}
\usage{
getOdinSpectrum(filename)
}
\arguments{
\item{filename}{file name including path of the binary file to read}
}
\value{
a list with components head, freq and data
}
\description{
Take a file name of a singular Odin scan in binary format and return a spectrum
(i.e. a list consisting of header, frequency and data vectors).
}
\examples{
\dontrun{
filenames <- list.files(datadir, patter="SPE")
L <- lapply(filenames, getOdinSpectrum)
class(L) <- "spectra"
H <- getHead(L)
print(H)
}
}
