% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getSubset}
\alias{getSubset}
\title{Get subset}
\usage{
getSubset(L, index)
}
\arguments{
\item{L}{a list of spectra (each with components 'head', 'freq' and 'data')}

\item{index}{an integer vector holding the indices of the spectra to retrieve}
}
\value{
the requested subset of spectra returned as a list
}
\description{
From a list of spectra, get the subset which is described by an index vector.
}

