% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getVelo}
\alias{getVelo}
\title{Get velocity vectors}
\usage{
getVelo(L)
}
\arguments{
\item{L}{a list of spectra, each with components 'head', 'freq' and 'data'}
}
\value{
a matrix having all the velcity vectors as columns
}
\description{
From a list of spectra, get the velocity vectors and return as matrix.
}
\seealso{
\code{\link{getFreq}}
}
