% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rdrp.R
\name{lineMask}
\alias{lineMask}
\title{Construct a line mask}
\usage{
lineMask(y, limits)
}
\arguments{
\item{y}{a matrix of frequency or velocity vectors, one per column}

\item{limits}{pairs of values which each define a window}
}
\description{
For a given set of frequency (or velocity) windows, construct a
matrix of logical values, defining the areas which should not be used
in baseline fitting.
}

