% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lines.spectrum}
\alias{lines.spectrum}
\title{Lines method for class spectrum}
\usage{
\method{lines}{spectrum}(x, type = "l", col = "red", ...)
}
\arguments{
\item{x}{a single spectrum}

\item{type}{the type of plot to be used, default is 'l', i.e. line}

\item{col}{the colour of the plot, default 'red'}

\item{...}{further arguments to be passed to generic lines function}
}
\description{
Draw a spectrum.
}
\examples{
\dontrun{
plot(S1, type='s')   # use 'histo' mode to plot spectrum
lines(S2, type='s')  # draw second spectrum on top
}
}

