% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeDataset.R
\name{makeDataset}
\alias{makeDataset}
\title{Make a dataset from a list of individual spectra.}
\usage{
makeDataset(L)
}
\arguments{
\item{L}{a list of lists with components head, freq, data}
}
\value{
a list with components head (data.frame), freq (matrix) and data (matrix)
}
\description{
Given a list of spectra (i.e. lists with head, freq and data components),
return a dataset (a list) which has a data.frame describing header information,
and two matrices (columns of frequency and data vectors).
}

