% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drp.R
\name{markRegion}
\alias{markRegion}
\title{Mark a region}
\usage{
markRegion(S, ...)
}
\arguments{
\item{S}{a single spectrum, the one that is currently plotted}

\item{...}{further parameters passed to 'identify'}
}
\value{
a logical vector, one per channel
}
\description{
Use the mouse to mark region(s) of a spectrum which should be avoided
in baseline fitting. The number of points to be clicked should be
even. The routine will call \code{mask(S, limits)} on exit, with limits
being the x-coordinates of all the clicked points.
}
\seealso{
\code{\link{mask}}
}
