% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mask}
\alias{mask}
\title{Construct a line mask}
\usage{
mask(S, limits)
}
\arguments{
\item{S}{a single spectrum}

\item{limits}{pairs of values, which each define a window}
}
\value{
a vector of logical values, one per channel
}
\description{
For a given set of frequency (or velocity) windows, construct a
matrix of logical values, defining the areas which should not be used
in baseline fitting.
}
