% Generated by roxygen2 (4.0.1): do not edit by hand
\name{modify}
\alias{modify}
\title{Modify a header column}
\usage{
modify(L, column, value)
}
\arguments{
\item{L}{a list of spectra}

\item{column}{a string specifying which header column to modify}

\item{value}{a vector holding the new values of the header column}
}
\description{
Supply new values for a column of the header
}

