% Generated by roxygen2 (4.0.1): do not edit by hand
\name{moment}
\alias{moment}
\title{Calculate moments}
\usage{
moment(S, mask)
}
\arguments{
\item{S}{a single spectrum}

\item{mask}{a logical vector equal to TRUE for all the channels that should
be integrated.}
}
\value{
minimum, maximum, mean, rms, skewness and kurtosis
}
\description{
Given a mask defining the spectral areas of interest, return the integrated
area over those region(s).
}

