% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeDataset.R
\name{pick}
\alias{pick}
\title{Pick a subset of spectra.}
\usage{
pick(ds, index, drop.levels = FALSE)
}
\arguments{
\item{ds}{the original dataset}

\item{index}{the rows (columns) to pick}
}
\value{
an object of the same kind as ds, but only keeping given rows (columns)
}
\description{
Given an integer vector of indices, construct a subset of a dataset by selecting
only the rows (columns) from the header (data) of the original dataset.
}

