% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spectra.R
\name{plot.spectra}
\alias{plot.spectra}
\title{Plot method for class spectra.}
\usage{
\method{plot}{spectra}(ds, index = seq(ncol(ds$data)), grid = NULL)
}
\arguments{
\item{ds}{a dataset}
}
\description{
Plot all spectra in a dataset using matplot.
}

