% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{plot.spectrum}
\alias{plot.spectrum}
\title{Plot method for class spectrum}
\usage{
\method{plot}{spectrum}(x, type = "l", col = "blue", ...)
}
\arguments{
\item{x}{a single spectrum}

\item{type}{the type of plot to be used, default is 'l', i.e. line}

\item{col}{the colour of the plot, default 'blue'}

\item{...}{further arguments to be passed to generic plot function}
}
\description{
Plot a single spectrum.
}
\examples{
\dontrun{
plot(S, type='s')   # use 'histo' mode to plot spectrum
}
}
