% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drp.R
\name{positions}
\alias{positions}
\title{Plot a set of positions}
\usage{
positions(lon, lat, ...)
}
\arguments{
\item{lon}{a numeric vector with longitudinal values, e.g. RA}

\item{lat}{a numeric vector with latitudinal values, e.g. Dec}

\item{...}{further arguments to be passed on to generic function plot}
}
\description{
Given two vectors lon and lat, which describe the longitudinal and
latitudinal coordinates of individual spectra, produce a scatter plot
taking into account projection effects.
}

