% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.spectra}
\alias{print.spectra}
\title{Print method for class spectra}
\usage{
\method{print}{spectra}(x, ...)
}
\arguments{
\item{x}{a list of spectra}

\item{...}{further arguments to be passed to generic function print}
}
\description{
Simply print the header (data.frame) part of the spectra.
}

