% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{print.spectrum}
\alias{print.spectrum}
\title{Print method for class spectrum}
\usage{
\method{print}{spectrum}(x, ...)
}
\arguments{
\item{x}{a single spectrum}

\item{...}{further arguments to be passed to generic function print}
}
\description{
Simply print the header part of the spectrum.
}

