% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataio.R
\name{readSALSA}
\alias{readSALSA}
\title{Turn FITS files into a list of spectra.}
\usage{
readSALSA(fitsfiles)
}
\arguments{
\item{fitsfiles}{file name(s) including path of FITS file(s) to convert}
}
\value{
a list of lists with components head, freq and data
}
\description{
Take a FITS file for a single spectrum from SALSA and return a spectrum
(i.e. a list consisting of header, frequency and data vectors).
}

