% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{resample}
\alias{resample}
\title{Resample spectrum}
\usage{
resample(S, f, smooth = FALSE)
}
\arguments{
\item{S}{a single spectrum}

\item{f}{a frequency vector onto which the spectrum should be resampled}

\item{smooth}{if TRUE convolve with Gaussian response}
}
\value{
the resampled spectrum
}
\description{
Perform resampling (cubic spline interpolation) on a single spectrum.
}

