% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rescale}
\alias{rescale}
\title{Rescale spectrum}
\usage{
rescale(S, factor = 1, bias = 0)
}
\arguments{
\item{S}{a single spectrum}

\item{factor}{a numeric value by which to scale the whole spectrum}

\item{bias}{a numeric value to add to all channels}
}
\value{
the rescaled spectrum, out = in * factor + bias
}
\description{
Add bias and/or scale by factor.
}

