% Generated by roxygen2 (4.0.1): do not edit by hand
\name{reverse}
\alias{reverse}
\title{Reverse a spectrum}
\usage{
reverse(S)
}
\arguments{
\item{S}{a single spectrum}
}
\value{
the reversed spectrum
}
\description{
Reverse the order of the channels, i.e. turn both the frequency and data
vector around.
}

