% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/drp.R
\name{stamp}
\alias{stamp}
\title{Plot a grid of spectra}
\usage{
stamp(L)
}
\arguments{
\item{L}{a list of spectra}
}
\description{
Try to find a grid of rectangles large enough to hold all the
spectra and as square like as possible, and "stamp" the individual
spectra into the grid cells.
}

