% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spectra.R
\name{summary.spectra}
\alias{summary.spectra}
\title{Summary method for class spectra.}
\usage{
\method{summary}{spectra}(ds, ...)
}
\arguments{
\item{ds}{a dataset}
}
\description{
Produce summary for all spectra in a dataset.
}

