% Generated by roxygen2 (4.0.1): do not edit by hand
\name{trim}
\alias{trim}
\title{Trim channels from spectra}
\usage{
trim(S, keep)
}
\arguments{
\item{S}{a single spectrum}

\item{keep}{a vector holding the channel numbers to keep}
}
\value{
the trimmed spectrum
}
\description{
Given a single spectrum and a vector of channel numbers, trim the spectrum such that
only the given channels are kept in the frequency and data vectors.
}

