% Generated by roxygen2 (4.0.1): do not edit by hand
\name{velocities}
\alias{velocities}
\title{Construct velocity vector}
\usage{
velocities(S)
}
\arguments{
\item{S}{a single spectrum}
}
\value{
a numeric vector holding velocities
}
\description{
Turn the frequency vector into a velocity vector.
}
\seealso{
\code{\link{frequencies}}
}

