% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rdrp.R
\name{velocity}
\alias{velocity}
\title{Calculate velocity vectors}
\usage{
velocity(f, f0, vs)
}
\arguments{
\item{f}{a matrix of frequency vectors for which velocities will be calculated}

\item{f0}{the rest frequency in the same units as f}

\item{vs}{the source velocity in km/s}
}
\value{
vel a matrix with one velocity vector per column
}
\description{
Given the frequency vectors and source velocities, calculate a matrix
of velocities corresponding to the ds$freq matrix.
}

