% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{addColumns}
\alias{addColumns}
\title{Add header columns}
\usage{
addColumns(L, newnames)
}
\arguments{
\item{L}{a list of spectra}

\item{newnames}{a character vector holding an array of column names to be
added to the header}
}
\description{
Add new columns to the header part of the spectra, you may fill these new
columns by using routine 'modify'.
}

