% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{area}
\alias{area}
\title{Calculate integrated area}
\usage{
area(S, mask)
}
\arguments{
\item{S}{a single spectrum}

\item{mask}{a logical vector equal to TRUE for all the channels that should
be integrated.}
}
\value{
the integrated value
}
\description{
Given a mask defining the spectral areas of interest, return the integrated
area over those region(s).
}
\examples{
data(salsa)
options(system="velocity")    # work in velocity space
S <- salsa[[1]]               # get the first spectrum
v <- velocities(S)            # get velocity vector
mask <- (v > -20) & (v < 20)  # integrate from -20..20 km/s
area(S, mask)                 # calculate integrated area in K*km/s
# call 'area' for each of the spectra in 'salsa' with parameter 'mask'
sapply(salsa, FUN=area, mask)
}
