% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rdrp.R
\name{average}
\alias{average}
\title{Average spectra.}
\usage{
average(ds, w = NA)
}
\arguments{
\item{ds}{dataset to be averaged}

\item{w}{weights to be used}
}
\value{
a dataset consisting of a single, averaged spectrum.
}
\description{
Average the spectra in the given dataset, using the vector of supplied weights.
If weights = NA, calculate weights based on Tsys and integration time.
}

