% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{average}
\alias{average}
\title{Calculate average spectrum}
\usage{
average(L)
}
\arguments{
\item{L}{a list of spectra (each with components 'head', 'freq' and 'data')}
}
\value{
the average spectrum
}
\description{
From a list of spectra, calculate the average spectrum where the weighting is
done using system temperature and integration times of the headers.
}

