% Generated by roxygen2 (4.0.1): do not edit by hand
\name{average}
\alias{average}
\title{Calculate average spectrum}
\usage{
average(L)
}
\arguments{
\item{L}{a list of spectra (each with components 'head', 'freq' and 'data')}
}
\value{
the average spectrum
}
\description{
From a list of spectra, calculate the average spectrum where the weighting is
done using system temperature and integration times of the headers.
}
\examples{
data(salsa)
plot(salsa)           # plot individual spectra in upper frame ...
A <- average(salsa)
lines(A$freq, A$data, lwd=5, col='red')
}

