% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{average}
\alias{average}
\title{Calculate average spectrum}
\usage{
average(L)
}
\arguments{
\item{L}{a list of spectra, each with components 'head', 'freq' and 'data'}
}
\value{
the average spectrum
}
\description{
From a list of spectra, calculate the average spectrum where the weighting is
done using system temperature and integration times of the headers.
}
\examples{
data(salsa)
A <- average(salsa)
plot(salsa)                              # plot individual spectra
lines(A$freq, A$data, lwd=5, col='red')  # draw average spectrum on top
}

