% Generated by roxygen2 (4.0.1): do not edit by hand
\name{baseline}
\alias{baseline}
\title{Fit a baseline to spectra in a dataset}
\usage{
baseline(S, order = 1, mask = NULL)
}
\arguments{
\item{S}{a single spectrum}

\item{order}{the order of the polynomial to fit}

\item{mask}{a channel line mask, fit baseline to channels where mask = FALSE}
}
\value{
the fitted baselines
}
\description{
For all spectra in a dataset fit a polynomial of given order, using only the
spectral channels given in window.
}
\examples{
data(salsa)
A <- average(salsa)
plot(A)
limits <- c(1419.9, 1421.1)
abline(v=limits, lty=2, col='grey')    # mark line region
linemask <- mask(A, limits)
bl <- baseline(A, order=3, mask=linemask)
lines(A$freq, bl, col='red')
}
\seealso{
\code{\link{mask}}
}

