% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rdrp.R
\name{baseline}
\alias{baseline}
\title{Fit a baseline to spectra in a dataset}
\usage{
baseline(ds, order = 1, window = NULL)
}
\arguments{
\item{ds}{a dataset}

\item{order}{the order of the polynomial to fit}

\item{window}{if given, a logical vector with TRUE for channels to be used}
}
\value{
a list of fitted baselines
}
\description{
For all spectra in a dataset fit a polynomial of given order, using only the
spectral channels given in window.
}

