% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drp.R
\name{baseline}
\alias{baseline}
\title{Fit a baseline to spectra in a dataset}
\usage{
baseline(S, order = 1, mask = NULL)
}
\arguments{
\item{S}{a single spectrum}

\item{order}{the order of the polynomial to fit}

\item{mask}{a channel line mask, fit baseline to channels where mask = FALSE}
}
\value{
a list giving the fitted baseline and the residuals of the fit
}
\description{
For all spectra in a dataset fit a polynomial of given order, using only the
spectral channels given in window.
}
\examples{
data(salsa)
A <- average(salsa)
limits <- c(1419.9, 1421.1)            # define line region
linemask <- mask(A, limits)            # turn into mask
bl <- baseline(A, order=3, mask=linemask)  # fit baseline
plot(A)                                # plot average spectrum
abline(v=limits, lty=2, col='grey')    # mark line region
print(bl$residuals)                    # print the residuals of the fit
lines(A$freq, bl$fitted, col='red')    # draw baseline on top
}
\seealso{
\code{\link{mask}}
}

