% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fold}
\alias{fold}
\title{Fold a frequency switched spectrum}
\usage{
fold(S, ft, shift = FALSE)
}
\arguments{
\item{S}{a single spectrum}

\item{ft}{frequency requency throw in Mhz}

\item{shift}{if TRUE, assume symmetric switching around true, centre frequency}
}
\value{
the folded spectrum
}
\description{
Fold a frequency switched spectrum
}

