% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getData}
\alias{getData}
\title{Get data vectors}
\usage{
getData(L)
}
\arguments{
\item{L}{a list of spectra, each with components 'head', 'freq' and 'data'}
}
\value{
a matrix having all the data vectors as columns
}
\description{
From a list of spectra, get the data vectors and return as matrix.
}
\examples{
data(salsa)
D <- getData(salsa)
image(D)             # show data matrix as color image
}
\seealso{
\code{\link{getFreq}}, \code{\link{getVelo}}
}

