% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getData}
\alias{getData}
\title{Get data vectors}
\usage{
getData(L)
}
\arguments{
\item{L}{a list of spectra (each with components 'head', 'freq' and 'data')}
}
\value{
a matrix having all the data vectors as columns
}
\description{
From a list of spectra, get the data vectors and return as matrix.
}

