% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getFreq}
\alias{getFreq}
\title{Get frequency vectors}
\usage{
getFreq(L)
}
\arguments{
\item{L}{a list of spectra, each with components 'head', 'freq' and 'data'}
}
\value{
a matrix having all the frequency vectors as columns
}
\description{
From a list of spectra, get the frequency vectors and return as matrix.
}
\seealso{
\code{\link{getVelo}}
}

