% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeDataset.R
\name{getHead}
\alias{getHead}
\title{Return headers of spectra.}
\usage{
getHead(L)
}
\arguments{
\item{L}{list of spectra}
}
\value{
a dataframe containing header info for each spectrum, one per row.
}
\description{
Turn head component of list of spectra into data.frame
}

