% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drp.R
\name{markRegion}
\alias{markRegion}
\title{Mark a region}
\usage{
markRegion(S, n = length(x), ...)
}
\arguments{
\item{S}{a single spectrum, the one that is currently plotted}

\item{n}{maximum number of points to be marked}

\item{...}{further parameters passed to 'identify'}
}
\value{
a logical vector, one per channel
}
\description{
Use the mouse to mark region(s) of a spectrum to be avoided in baseline
fitting. The number of points to be clicked should be even. The logical
vector returned is of the same type as if mask(S, limits) hade been called
with limits being the x-coordinates of all the clicked points.
}
\seealso{
mask
}

