% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drp.R
\name{modify}
\alias{modify}
\title{Modify a header column}
\usage{
modify(L, colname, value)
}
\arguments{
\item{L}{a list of spectra}

\item{colname}{a string specifying which header column to modify}

\item{value}{a vector holding the new values of the header column}
}
\description{
Supply new values for a column of the header
}

