% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{moment}
\alias{moment}
\title{Calculate moments}
\usage{
moment(S, mask)
}
\arguments{
\item{S}{a single spectrum}

\item{mask}{a logical vector equal to TRUE for all the channels that should
be integrated.}
}
\value{
a data frame containing minimum, maximum, mean, sdev, skewness and kurtosis
}
\description{
Given a mask defining the spectral areas of interest, return the integrated
area over those region(s).
}
\examples{
data(salsa)
A <- average(salsa)
lmask <- mask(A, c(1420,1421))  # calculate moments between 1420...1421 MHz
moment(A, mask=lmask)
do.call("rbind", lapply(salsa, moment, lmask)) # do this for individual spectra
}
\seealso{
\code{\link{area}}
}

