% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spectra.R
\name{plot.spectra}
\alias{plot.spectra}
\title{Plot method for class spectra}
\usage{
\method{plot}{spectra}(x, type = "l", col = "blue", ..., grid = NULL)
}
\arguments{
\item{x}{a list of spectra}

\item{type}{the type of plot to be used, default is 'l', i.e. line}

\item{col}{the colour of the plot, default 'blue'}

\item{...}{further arguments to be passed to generic function plot}

\item{grid}{specify grid (nrow x ncol) on which to organize spectra}
}
\description{
Plot all spectra in a dataset using matplot or a grid of plots.
}

