% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.spectrum}
\alias{plot.spectrum}
\title{Plot method for class spectrum}
\usage{
\method{plot}{spectrum}(x, type = "l", col = "blue", ...)
}
\arguments{
\item{x}{a single spectrum}

\item{type}{the type of plot to be used, default is 'l', i.e. line}

\item{col}{the colour of the plot, default 'blue'}

\item{...}{further arguments to be passed to generic function plot}
}
\description{
Plot a single spectrum.
}

