% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spectra.R
\name{print.spectra}
\alias{print.spectra}
\title{Print method for class spectra.}
\usage{
\method{print}{spectra}(ds, ...)
}
\arguments{
\item{ds}{a dataset}
}
\description{
Simply print the header (data.frame) part of the spectra.
}

