% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readClass}
\alias{readClass}
\title{Read a GILDAS/CLASS single dish data file}
\usage{
readClass(filename)
}
\arguments{
\item{filename}{name of the GILDAS file including path to be opened}
}
\value{
list of length n, where n is the number of scans found
}
\description{
Given a file name, open the file and scan it for single dish spectra
or continuum scans. Each encountered scan is returned as a list with
a head, freq and data section. For continuum scans, the frequency
vector will simply be a running index. All the individual lists are
combined into one major list, which is returned.
}

