% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readOSO20m}
\alias{readOSO20m}
\title{Read FITS files from the Onsala 20m telescope.}
\usage{
readOSO20m(fitsfiles)
}
\arguments{
\item{fitsfiles}{file name(s) including path of FITS file(s) to convert}
}
\value{
a list of lists with components head, freq and data
}
\description{
Take a FITS file for a single spectrum from the OSO 20m and return a spectrum
(i.e. a list consisting of header, frequency and data vectors).
}

