% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readSALSA}
\alias{readSALSA}
\title{Read a FITS file from the Onsala SALSA student telescope.}
\usage{
readSALSA(fitsfiles)
}
\arguments{
\item{fitsfiles}{file name(s) including path of FITS file(s) to convert}
}
\value{
a list of lists with components head, freq and data
}
\description{
Take a FITS file for a single spectrum from SALSA and return a spectrum
(i.e. a list consisting of header, frequency and data vectors).
}

