% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{salsa}
\alias{salsa}
\title{Some 21cm spectra taken with the Onsala SALSA student telescope.}
\format{A list with 29 elements, each of class "spectrum"}
\usage{
salsa
}
\description{
Spectra at 1421 MHz towards various galactic longitudes.
}
\keyword{datasets}

