% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sieve}
\alias{sieve}
\title{Filter spectrum}
\usage{
sieve(S, coeffs)
}
\arguments{
\item{S}{a single spectrum}

\item{coeffs}{a numeric vector with an odd number of filter coefficients}
}
\value{
the filtered spectrum
}
\description{
Perform filtering on a single spectrum.
}

