% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{trim}
\alias{trim}
\title{Trim channels from spectra}
\usage{
trim(S, keep)
}
\arguments{
\item{S}{a single spectrum}

\item{keep}{a vector holding the channel numbers to keep}
}
\value{
the trimmed spectrum
}
\description{
Given a single spectrum and a vector of channel numbers, trim the spectrum such that
only the given channels are kept in the frequency and data vectors.
}

