% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{velocities}
\alias{velocities}
\title{Construct velocity vector}
\usage{
velocities(S)
}
\arguments{
\item{S}{a single spectrum}
}
\value{
a numeric vector holding velocities
}
\description{
Turn the frequency vector into a velocity vector.
}
\seealso{
\code{\link{frequencies}}
}
