% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{velocity}
\alias{velocity}
\title{Construct velocity vector}
\usage{
velocity(S)
}
\arguments{
\item{S}{a single spectrum}
}
\value{
a numeric vector holding velocities
}
\description{
Turn the frequency vector into a velocity vector.
}

